AlarmConfigForm updated for BootTrigger - waiting for further implementation.
Removed option to choose between all triggers. BootTrigger and TimeTrigger both remain untouched in that aspect.

Solved issue:
Thought SQL queries did not reach the server. Though it did reach the server. The issue lied in the naming convention. Apparently "Trigger" is already specified and cannot or rather must NOT be used as a name in the table structure.

New Issue:
Consult Screenshot ("BackgroundTask Issue 2024-08-21 1952")