Added new class "BackgroundProcess" which does not show itself in the main taskbar but rather, on OS Windows, in the hidden symbols section. After closing the main thread, the task manager still shows the "SMSAlertSys" app running due to not having closed all threads (referring to screenshot "taskmanager second thread 2024-08-24 1920"). This new class is wholly based on "https://anujyadavcse.blogspot.com/2012/01/create-background-taskbar-application.html".
Added the instantiation of new thread which runs a new application (instantiates the freshly implemented class "BackgroundProcess")

Added button for new thread.

------------

Further development of this new class is needed. It should, in the end, compare the current date time and the ones that the user had put in. Also found out what "static" does now with its usage being demonstrated in the integration of a custom icon ("exampleIcon").

Further development needed: XML file in which the dates are saved when one does not have a SQL server or does not have access to one.